now its working, changed the method for subdivisions
so the spiral subdisions added for consistent shape when the radius and arc increases as per formula, arc = radius * angle. large circles have more area...
it will divide the area into multiple parts and draws spiral on each of them instead of just one which was unable to cover the large arc.

DESIRED_SPIRAL_ARC constant defines that.