water brush tested
water brush added,
with a grid mesh generatedd on CPU and its vertices being displaced from heightmap.
water currenty stick to the surface, it will need dynamic current flow map to be able to show a water wall  effect.

the current water mesh generatation of 1024* 1024 heightmap, water map is currently using the same size is slow and gives a spike in fps when the mesh is generated after user stops clicking.

1024*1024 = 1,048,576
512 * 512 = 262,144

its 4x difference after each step.

so now i will make the dynamic currrentmap generation, then will move to rewrite and its optimization.

the mesh generation of water can be multithreaded and its generation can be divided into multiple frames instead of one. or it can be generated on GPU.
one upper hand on GPU is that the heightmap is present on gpu, i dont want to use compute pipline as it may stag(add wait time for) the graphics pipeline.

and the water working design is needed.